+++ splinter/driver/webdriver/__init__.py
@@ -3,6 +3,7 @@
 import time
 import logging
 import subprocess
+from contextlib import contextmanager
 
 from tempfile import TemporaryFile
 from lxml.cssselect import CSSSelector
@@ -11,8 +12,8 @@ from selenium.webdriver.firefox import firefox_profile
 
 from splinter.driver import DriverAPI, ElementAPI
 from splinter.element_list import ElementList
+from splinter.utils import warn_deprecated
 
-import time
 
 class BaseWebDriver(DriverAPI):
     old_popen = subprocess.Popen
@@ -73,6 +74,9 @@ class BaseWebDriver(DriverAPI):
     def visit(self, url):
         self.driver.get(url)
 
+    def reload(self):
+        self.driver.refresh()
+
     def execute_script(self, script):
         self.driver.execute_script(script)
 
@@ -97,11 +101,15 @@ class BaseWebDriver(DriverAPI):
                 return True
         return False
 
-    def is_element_present_by_css_selector(self, css_selector, wait_time=None):
-        return self.is_element_present(self.find_by_css_selector, css_selector, wait_time)
+    def is_element_present_by_css(self, css_selector, wait_time=None):
+        return self.is_element_present(self.find_by_css, css_selector, wait_time)
 
-    def is_element_not_present_by_css_selector(self, css_selector):
-        return self.is_element_not_present(self.find_by_css_selector, css_selector)
+    is_element_present_by_css_selector = warn_deprecated(is_element_present_by_css, 'is_element_present_by_css_selector')
+
+    def is_element_not_present_by_css(self, css_selector):
+        return self.is_element_not_present(self.find_by_css, css_selector)
+
+    is_element_not_present_by_css_selector = warn_deprecated(is_element_not_present_by_css, 'is_element_not_present_by_css_selector')
 
     def is_element_present_by_xpath(self, xpath, wait_time=None):
         return self.is_element_present(self.find_by_xpath, xpath, wait_time)
@@ -113,7 +121,7 @@ class BaseWebDriver(DriverAPI):
         return self.is_element_present(self.find_by_tag, tag, wait_time)
 
     def is_element_not_present_by_tag(self, tag):
-        return self.is_element_present(self.find_by_tag, tag)
+        return self.is_element_not_present(self.find_by_tag, tag)
 
     def is_element_present_by_name(self, name, wait_time=None):
         return self.is_element_present(self.find_by_name, name, wait_time)
@@ -122,29 +130,21 @@ class BaseWebDriver(DriverAPI):
         return self.is_element_not_present(self.find_by_name, name, wait_time)
 
     def is_element_present_by_id(self, id, wait_time=None):
-        wait_time = wait_time or self.wait_time
-        end_time = time.time() + wait_time
-
-        while time.time() < end_time:
-            try:
-                self.driver.find_element_by_id(id)
-                return True
-            except NoSuchElementException:
-                pass
-        return False
+        return self.is_element_present(self.find_by_id, id, wait_time)
 
     def is_element_not_present_by_id(self, id):
-        end_time = time.time() + self.wait_time
-
-        while time.time() < end_time:
-            try:
-                self.driver.find_element_by_id(id)
-            except NoSuchElementException:
-                return True
-        return False
-
-    def switch_to_frame(self, id):
+        return self.is_element_not_present(self.find_by_id, id)
+        
+    def get_alert(self):
+        return AlertElement(self.driver.switch_to_alert())
+    
+    @contextmanager
+    def get_iframe(self, id):
         self.driver.switch_to_frame(id)
+        try:
+            yield self
+        finally:
+            self.driver.switch_to_frame(None)
 
     def find_option_by_value(self, value):
         return self.find_by_xpath('//option[@value="%s"]' % value)
@@ -156,63 +156,47 @@ class BaseWebDriver(DriverAPI):
         return self.find_by_xpath('//a[@href="%s"]' % href)
 
     def find_link_by_text(self, text):
-        return ElementList([self.element_class(element) for element in self.driver.find_elements_by_link_text(text)])
+        return ElementList([self.element_class(element, self) for element in self.driver.find_elements_by_link_text(text)])
 
-    def find_by_css_selector(self, css_selector):
-        selector = CSSSelector(css_selector)
-
-        end_time = time.time() + self.wait_time
-
-        while time.time() < end_time:
-            elements = self.driver.find_elements_by_xpath(selector.path)
-            if elements:
-                return ElementList([self.element_class(element) for element in elements])
-        return ElementList([])
-
-    def find_by_xpath(self, xpath):
-        end_time = time.time() + self.wait_time
-
-        while time.time() < end_time:
-            elements = self.driver.find_elements_by_xpath(xpath)
-            if elements:
-                return ElementList([self.element_class(element) for element in elements])
-        return ElementList([])
-
-    def find_by_name(self, name):
-        end_time = time.time() + self.wait_time
-
-        while time.time() < end_time:
-            elements = self.driver.find_elements_by_name(name)
-            if elements:
-                return ElementList([self.element_class(element) for element in elements])
-        return ElementList([])
-
-    def find_by_id(self, id):
+    def find_by(self, finder, selector):
+        elements = None
         end_time = time.time() + self.wait_time
 
         while time.time() < end_time:
             try:
-                element = self.driver.find_element_by_id(id)
-                return ElementList([self.element_class(element)])
+                elements = finder(selector)
+                if not isinstance(elements, list):
+                    elements = [elements]
             except NoSuchElementException:
                 pass
 
+            if elements:
+                return ElementList([self.element_class(element, self) for element in elements])
         return ElementList([])
+
+    def find_by_css(self, css_selector):
+        selector = CSSSelector(css_selector)
+        return self.find_by(self.driver.find_elements_by_xpath, selector.path)
+
+    find_by_css_selector = warn_deprecated(find_by_css, 'find_by_css_selector')
+
+    def find_by_xpath(self, xpath):
+        return self.find_by(self.driver.find_elements_by_xpath, xpath)
+
+    def find_by_name(self, name):
+        return self.find_by(self.driver.find_elements_by_name, name)
 
     def find_by_tag(self, tag):
-        end_time = time.time() + self.wait_time
+        return self.find_by(self.driver.find_elements_by_tag_name, tag)
 
-        while time.time() < end_time:
-            elements = self.driver.find_elements_by_tag_name(tag)
-            if elements:
-                return ElementList([self.element_class(element) for element in elements])
-        return ElementList([])
+    def find_by_id(self, id):
+        return self.find_by(self.driver.find_element_by_id, id)
 
-    def fill_in(self, name, value):
+    def fill(self, name, value):
         field = self.find_by_name(name).first
         field.value = value
 
-    fill = fill_in
+    fill_in = warn_deprecated(fill, 'fill_in')
     attach_file = fill
 
     def choose(self, name):
@@ -236,8 +220,9 @@ class BaseWebDriver(DriverAPI):
 
 class WebDriverElement(ElementAPI):
 
-    def __init__(self, element):
+    def __init__(self, element, parent):
         self._element = element
+        self.parent = parent
 
     def _get_value(self):
         try:
@@ -278,3 +263,25 @@ class WebDriverElement(ElementAPI):
 
     def __getitem__(self, attr):
         return self._element.get_attribute(attr)
+
+
+class AlertElement(object):
+    
+    def __init__(self, alert):
+        self._alert = alert
+        self.text = alert.text
+        
+    def accept(self):
+        self._alert.accept()
+        
+    def dismiss(self):
+        self._alert.dismiss()
+    
+    def fill_with(self, text):
+        self._alert.send_keys(text)
+        
+    def __enter__(self):
+        return self
+        
+    def __exit__(self, type, value, traceback):
+        pass