+++ tests/iframes.py
@@ -2,13 +2,10 @@ from nose.tools import assert_equals
 
 class IFrameElementsTest(object):
 
-    def test_can_change_frames(self):
-        self.browser.switch_to_frame('iframemodal')
-        value = self.browser.find_by_tag('h1').first.value
-        assert_equals(value, 'IFrame Example Header')
-
-    def test_can_change_frames_back(self):
-        self.browser.switch_to_frame('iframemodal')
-        self.browser.switch_to_frame(None)
+    def test_can_work_on_iframes(self):
+        """can work on iframes and switch back to the page"""
+        with self.browser.get_iframe('iframemodal') as frame:
+            value = frame.find_by_tag('h1').first.value
+            assert_equals(value, 'IFrame Example Header')
         value = self.browser.find_by_tag('h1').first.value
         assert_equals(value, 'Example Header')